modified demoprep.sh in .

+++ demoprep.sh
@@ -108,7 +108,9 @@ brew install neovim
 rm -rf ~/.config/nvim
 git clone https://github.com/AstroNvim/AstroNvim ~/.config/nvim
 
-wget https://www.dropbox.com/s/e6r7awhvai3nexx/VMware-Fusion-e.x.p-20486664_universal.dmg?dl=1 -O VMware.dmg
+if [ ! -f VMware.dmg ]; then
+    wget https://www.dropbox.com/s/e6r7awhvai3nexx/VMware-Fusion-e.x.p-20486664_universal.dmg?dl=1 -O VMware.dmg
+fi
 
 CURRENTDOCK="$(dockutil -L | cut -f1)"
